Added tests for automatic differentiation (2/3 broken)
