+++ test/specific_problems.jl
@@ -6,7 +6,6 @@ using Test
 using LinearAlgebra
 using UnPack
 using ParameterizedFunctions
-using ForwardDiff
 
 
 using DiffEqProblemLibrary.ODEProblemLibrary: importodeproblems; importodeproblems()
@@ -135,22 +134,4 @@ end
     u0 = [1.0,1.0]
     prob = ODEProblem(f,u0,tspan,p)
     @test solve(prob, EK1(order=3)) isa ProbNumDiffEq.ProbODESolution
-end
-
-
-@testset "Derivative with ForwardDiff.jl" begin
-    prob = ProbNumDiffEq.remake_prob_with_jac(prob_ode_fitzhughnagumo)
-    function param_to_loss(p)
-        sol = solve(remake(prob, p=p), EK1(order=3))
-        return norm(sol.u[end])  # Dummy loss
-    end
-    @test ForwardDiff.gradient(param_to_loss, prob.p) ≈
-        [0.026680212891877435, -0.028019989130281753, 0.3169977494388167, 0.6749351039218744]
-
-    function startval_to_loss(u0)
-        sol = solve(remake(prob, u0=u0), EK1(order=3))
-        return norm(sol.u[end])  # Dummy loss
-    end
-    @test ForwardDiff.gradient(startval_to_loss, prob.u0) ≈
-        [0.6500925873857853, -0.004812245513746423]
 end
